Add files via upload
Here's how the code works in plain language:

1. Create the Main Window:

We start by creating the main window, or JFrame, which is like the main screen of your application. This is where everything will be displayed.
2. Set Up a CardLayout:

We set up a special layout manager called CardLayout. This layout manager allows us to easily switch between different "screens" within the main window.
3. Create Panels for Different Screens:

We create separate panels to represent the different screens or pages of our application. These panels are like different views you want to show to the user. In your case, you have panels for "Home," "Expenses," and "Budget."
4. Design the Content of Each Panel:

We design what should be displayed on each panel. For example, on the "Home" panel, we show a welcome message. On the "Expenses" panel, we have a message and an input field. On the "Budget" panel, we show a message about savings.
5. Add Panels to the Main Panel:

We add these panels to a main panel that uses the CardLayout. This main panel will be the container that holds all the different screens.
6. Create Buttons for Navigation:

We create buttons that act as a menu. Clicking on these buttons allows us to switch between the different screens.
7. Attach Action Listeners to Buttons:

We make the buttons responsive by attaching action listeners. When a button is clicked, it tells the program to show a specific panel.
8. Create a Menu Panel:

We create another panel that holds the menu buttons. This panel is placed at the bottom of the main window.
9. Add Panels to the Main Window:

We add the main panel with the CardLayout and the menu panel to the main window.
10. Display the Main Window:
- Finally, we set some properties for the main window and make it visible to the user.

The key idea here is that the CardLayout lets you switch between different panels without recreating them each time. You can think of the main panel as a stack of cards, and the CardLayout helps you flip through these cards to show the one you want. The menu buttons are like tabs that let you navigate through the cards.